# docs/swagger.yaml
schemes:
  - http
  - https
swagger: '2.0'
info:
  description: REST API for the LMS platform covering auth, courses, enrollments, assessments, payments, and more.
  title: LMS Backend API
  contact: {}
  version: 1.0.0
host: localhost:8080
basePath: /
paths:
...
  /v1/auth/admin/login:
    post:
      description: Authenticates an admin user with username and password; returns an OTP session token for two-factor verification
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - auth
      summary: Admin login
      parameters:
        - description: Admin login request
          name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              password:
                type: string
              username:
                type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/auth.OTPSessionResult'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/auth/admin/resend-otp:
    post:
      description: Resends the OTP for an active admin login session identified by the OTP session token
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - auth
      summary: Admin resend OTP
      parameters:
        - description: Admin resend OTP request
          name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              otp_session_token:
                type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              message:
                type: string
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '429':
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
  /v1/auth/admin/verify-otp:
    post:
      description: Verifies the admin's OTP using the session token from AdminLogin and returns JWT tokens on success
      consumes:
        - application/json
      produces:
        - application/json
      tags:
        - auth
      summary: Admin verify OTP
      parameters:
        - description: Admin OTP verification request
          name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              otp:
                type: string
              otp_session_token:
                type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/auth.TokenPair'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
...
  /v1/certificates/verify/{verificationId}:
    get:
      description: Public endpoint that verifies a certificate by its verification ID and returns its validity and public details
      produces:
        - application/json
      tags:
        - certificates
      summary: Verify certificate
      parameters:
        - type: string
          description: Verification ID
          name: verificationId
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/certificates.VerifyCertificateResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
...
  /v1/student/certificates/{courseId}:
    get:
      security:
        - BearerAuth: []
      description: Returns the authenticated student's certificate for a completed course, including a presigned PDF URL
      produces:
        - application/json
      tags:
        - certificates
      summary: Get student certificate
      parameters:
        - type: string
          description: Course ID
          name: courseId
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/certificates.CertificateResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ValidationErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '403':
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
...
definitions:
...
  auth.OTPSessionResult:
    type: object
    properties:
      expires_at:
        type: string
      masked_email:
        type: string
      otp_session_token:
        type: string
...
  auth.TokenPair:
    type: object
    properties:
      access_token:
        type: string
      expires_in:
        description: seconds
        type: integer
      refresh_token:
        type: string
...
  certificates.CertificateResponse:
    type: object
    properties:
      completion_date:
        description: ISO 8601 date string
        type: string
      course_id:
        type: string
      course_title:
        type: string
      created_at:
        type: string
      id:
        type: string
      instructor_name:
        type: string
      pdf_url:
        description: presigned URL, nil if PDF not yet generated
        type: string
      student_id:
        type: string
      student_name:
        type: string
      verification_id:
        type: string
  certificates.VerifyCertificateResponse:
    type: object
    properties:
      completion_date:
        description: ISO 8601 date string
        type: string
      course_title:
        type: string
      instructor_name:
        type: string
      issued_at:
        type: string
      student_name:
        type: string
      valid:
        type: boolean
      verification_id:
        type: string
...
  handlers.ErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          code:
            type: string
          message:
            type: string
  handlers.ValidationDetail:
    type: object
    properties:
      field:
        type: string
      message:
        type: string
  handlers.ValidationErrorResponse:
    type: object
    properties:
      error:
        type: object
        properties:
          code:
            type: string
          details:
            type: array
            items:
              $ref: '#/definitions/handlers.ValidationDetail'
          message:
            type: string
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header